Refer to project as "F# Charting" in documentation
This commit modifies the headers as well as the text of the
documentation to refer to the project as "F# Charting" as opposed
to "FSharp.Charting," except when referring to the NuGet package,
namespace, assembly, etc. This is the convention followed by the
F# Data documenation as well.

To avoid redundancy, I also changed the home page title to
"F# Charting: Library for Data Visualization" as opposed to
"F# Charting: The F# Charting Library." Feel free to change if you
think that doesn't work or come up with a better title.