Add relevant include directories for Ubuntu
In Ubuntu and the likes gtk-sharp, atk-sharp, gdk-sharp and glib-sharp are split into different directories. Thus it would help to have these automatically included in the script. It is also all important that the libglibsharpglue-2.so and appropriate .config files can be found on the path. The active includes should not hurt, even, if the directories are not present, but it is OK also to comment them out.