Combine extra margins for legend and title instead of overwriting
When a chart has a title and a legend, and both are docked outside the
chart area, the extra margin required to display the legend is ignored
because the extra margin for the title overwrites it. As a result, the
legend is not visible.

Rewrite the `computeExtraDefaultMargins` function to combine (by picking
the maximum ones) the extra margins required to display a legend and a
title. This makes the legend and the title visible in the chart even when
both are enabled.

Fixes #133